Refactor optimizer rules
