+++ arangod/Aql/Optimizer/Rule/JoinIndexNodes.h
@@ -24,19 +24,12 @@
 
 #pragma once
 
-#include "Aql/ExecutionNode/ExecutionNode.h"
 #include "Aql/ExecutionPlan.h"
-#include "Aql/OptimizerRulesFeature.h"
 
 namespace arangodb::aql {
 class Optimizer;
-class SubqueryNode;
 
-class QueryContext;
-struct Collection;
-
-/// @brief split and-combined filters and break them into smaller parts
-void splitFiltersRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
-                      OptimizerRule const&);
+void joinIndexNodesRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
+                        OptimizerRule const&);
 
 }  // namespace arangodb::aql
