+++ arangod/Aql/Optimizer/Rule/PushFilterIntoEnumerateNear.cpp
@@ -21,6 +21,8 @@
 /// @author Jure Bajic
 ////////////////////////////////////////////////////////////////////////////////
 
+#include "PushFilterIntoEnumerateNear.h"
+
 #include "Aql/Ast.h"
 #include "Aql/AstHelper.h"
 #include "Aql/Collection.h"
@@ -34,8 +36,7 @@
 #include "Aql/ExecutionNode/FilterNode.h"
 #include "Aql/ExecutionNode/SortNode.h"
 #include "Aql/Optimizer.h"
-#include "Aql/Optimizer/Rule/OptimizerRuleVectorIndexHelpers.h"
-#include "Aql/OptimizerRules.h"
+#include "Aql/Optimizer/Utils/VectorIndexHelpers.h"
 #include "Aql/OptimizerUtils.h"
 #include "Aql/QueryContext.h"
 #include "Basics/Exceptions.h"
