+++ arangod/Aql/Optimizer/Rule/SubstituteClusterMultipleDocumentOperations.h
@@ -24,19 +24,13 @@
 
 #pragma once
 
-#include "Aql/ExecutionNode/ExecutionNode.h"
 #include "Aql/ExecutionPlan.h"
-#include "Aql/OptimizerRulesFeature.h"
 
 namespace arangodb::aql {
 class Optimizer;
-class SubqueryNode;
 
-class QueryContext;
-struct Collection;
-
-/// @brief split and-combined filters and break them into smaller parts
-void splitFiltersRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
-                      OptimizerRule const&);
+/// @brief replace multiple document operations in cluster by special handling
+void substituteClusterMultipleDocumentOperationsRule(
+    Optimizer* opt, std::unique_ptr<ExecutionPlan> plan, OptimizerRule const&);
 
 }  // namespace arangodb::aql
