+++ arangod/Aql/Optimizer/Utils/CMakeLists.txt
@@ -1,4 +1,8 @@
 target_sources(arango_aql PRIVATE
   AccessesCollectionVariable.cpp
+  DepIsSingletonOrConstCalc.cpp
+  GetAstNode.cpp
+  GetFunction.cpp
   OptimizeTraversalPathVariable.cpp
+  ReplaceNode.cpp
 )
