+++ arangod/Aql/Optimizer/Utils/GetFunction.h
@@ -18,25 +18,15 @@
 ///
 /// Copyright holder is ArangoDB GmbH, Cologne, Germany
 ///
-/// @author Max Neunhoeffer
-/// @author Jan Steemann
+/// @author Jan Christoph Uhde
 ////////////////////////////////////////////////////////////////////////////////
 
 #pragma once
 
-#include "Aql/ExecutionNode/ExecutionNode.h"
-#include "Aql/ExecutionPlan.h"
-#include "Aql/OptimizerRulesFeature.h"
-
 namespace arangodb::aql {
-class Optimizer;
-class SubqueryNode;
+struct AstNode;
+struct Function;
 
-class QueryContext;
-struct Collection;
-
-/// @brief split and-combined filters and break them into smaller parts
-void splitFiltersRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
-                      OptimizerRule const&);
+Function* getFunction(AstNode const* ast) noexcept;
 
 }  // namespace arangodb::aql
