+++ arangod/Aql/Optimizer/Utils/ReplaceNode.h
@@ -18,25 +18,17 @@
 ///
 /// Copyright holder is ArangoDB GmbH, Cologne, Germany
 ///
-/// @author Max Neunhoeffer
-/// @author Jan Steemann
+/// @author Wilfried Goesgens
+/// @author Jan Christoph Uhde
 ////////////////////////////////////////////////////////////////////////////////
 
 #pragma once
 
-#include "Aql/ExecutionNode/ExecutionNode.h"
-#include "Aql/ExecutionPlan.h"
-#include "Aql/OptimizerRulesFeature.h"
-
 namespace arangodb::aql {
-class Optimizer;
-class SubqueryNode;
+class ExecutionNode;
+class ExecutionPlan;
 
-class QueryContext;
-struct Collection;
-
-/// @brief split and-combined filters and break them into smaller parts
-void splitFiltersRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
-                      OptimizerRule const&);
+void replaceNode(ExecutionPlan* plan, ExecutionNode* oldNode,
+                 ExecutionNode* newNode);
 
 }  // namespace arangodb::aql
