+++ arangod/Aql/Optimizer/Utils/VectorIndexHelpers.h
