+++ arangod/Aql/OptimizerRulesFeature.cpp
@@ -75,8 +75,15 @@
 #include "Aql/Optimizer/Rule/PushDownLateMaterialization.h"
 #include "Aql/Optimizer/Rule/MaterializeIntoSeparateVariable.h"
 #include "Aql/Optimizer/Rule/PushLimitIntoIndex.h"
+#include "Aql/Optimizer/Rule/ReplaceLikeWithRange.h"
 #include "Aql/Optimizer/Rule/ReplaceEntriesWithObjectIteration.h"
 #include "Aql/Optimizer/Rule/ReplaceEqualAttributeAccess.h"
+#include "Aql/Optimizer/Rule/ReplaceNearWithinFulltext.h"
+#include "Aql/Optimizer/Rule/SubstituteClusterSingleDocumentOperations.h"
+#include "Aql/Optimizer/Rule/SubstituteClusterMultipleDocumentOperations.h"
+#include "Aql/Optimizer/Rule/JoinIndexNodes.h"
+#include "Aql/Optimizer/Rule/UseVectorIndex.h"
+#include "Aql/Optimizer/Rule/PushFilterIntoEnumerateNear.h"
 #include "Aql/OptimizerRules.h"
 #include "Aql/OptimizerRulesOptionsProvider.h"
 #ifdef USE_ENTERPRISE
